+++ log4j2.xml
@@ -9,13 +9,9 @@
 
     <Loggers>
 
-        <Logger name="com.aayushatharva.sourcecenginequerycacher" level="debug">
+        <Logger name="com.aayushatharva.seqc" level="debug">
             <AppenderRef ref="ConsoleAppender"/>
         </Logger>
-
-        <Root level="debug">
-                        <AppenderRef ref="ConsoleAppender"/>
-        </Root>
     </Loggers>
 
 </Configuration>
