+++ pom.xml
@@ -17,17 +17,17 @@
 
         <!-- Version Management -->
         <commons_cli>1.4</commons_cli>
-        <log4j2>2.15.0</log4j2>
+        <log4j2>2.16.0</log4j2>
         <junit>5.6.2</junit>
         <surefire>2.22.1</surefire>
     </properties>
 
     <dependencies>
-        <!-- https://mvnrepository.com/artifact/io.netty/netty-handler -->
+        <!-- https://mvnrepository.com/artifact/io.netty/netty-transport-native-epoll -->
         <dependency>
             <groupId>io.netty</groupId>
-            <artifactId>netty-all</artifactId>
-            <version>4.1.71.Final</version>
+            <artifactId>netty-transport-native-epoll</artifactId>
+            <version>4.1.72.Final</version>
         </dependency>
 
         <!-- https://mvnrepository.com/artifact/io.netty.incubator/netty-incubator-transport-native-io_uring -->
@@ -96,7 +96,7 @@
                         <configuration>
                             <transformers>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>com.aayushatharva.sourcecenginequerycacher.Main</mainClass>
+                                    <mainClass>com.aayushatharva.seqc.Main</mainClass>
                                     <manifestEntries>
                                         <Multi-Release>true</Multi-Release>
                                     </manifestEntries>
