+++ src/test/java/com/aayushatharva/sourcecenginequerycacher/HandlerTest.java
@@ -1,7 +1,8 @@
 package com.aayushatharva.sourcecenginequerycacher;
 
-import com.aayushatharva.sourcecenginequerycacher.utils.Config;
-import com.aayushatharva.sourcecenginequerycacher.utils.Packets;
+import com.aayushatharva.seqc.Main;
+import com.aayushatharva.seqc.utils.Config;
+import com.aayushatharva.seqc.utils.Packets;
 import io.netty.buffer.ByteBufUtil;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
@@ -45,7 +46,7 @@ class HandlerTest {
     @Order(1)
     void A2SInfoChallenge() throws IOException {
         DatagramPacket queryPck = new DatagramPacket(ByteBufUtil.getBytes(Packets.A2S_INFO_REQUEST), 0,
-                Packets.A2S_INFO_REQUEST.readableBytes(), Config.LocalServer.getAddress(), Config.LocalServer.getPort());
+                Packets.A2S_INFO_REQUEST.readableBytes(), Config.ServerAddress.getAddress(), Config.ServerAddress.getPort());
         byte[] responseBytes = new byte[4096];
         DatagramPacket responsePacket = new DatagramPacket(responseBytes, responseBytes.length);
 
@@ -66,7 +67,7 @@ class HandlerTest {
     void A2SInfo() throws IOException {
         byte[] Response = joinArrays(ByteBufUtil.getBytes(Packets.A2S_INFO_REQUEST), a2sChallenge);
         DatagramPacket queryPck = new DatagramPacket(Response, 0, Response.length,
-                Config.LocalServer.getAddress(), Config.LocalServer.getPort());
+                Config.ServerAddress.getAddress(), Config.ServerAddress.getPort());
 
         byte[] responseBytes = new byte[4096];
         DatagramPacket responsePacket = new DatagramPacket(responseBytes, responseBytes.length);
@@ -85,7 +86,7 @@ class HandlerTest {
     @Order(3)
     void A2SPlayerChallenge() throws IOException {
         DatagramPacket queryPck = new DatagramPacket(ByteBufUtil.getBytes(Packets.A2S_PLAYER_CHALLENGE_REQUEST_1), 0,
-                ByteBufUtil.getBytes(Packets.A2S_PLAYER_CHALLENGE_REQUEST_1).length, Config.LocalServer.getAddress(), Config.LocalServer.getPort());
+                ByteBufUtil.getBytes(Packets.A2S_PLAYER_CHALLENGE_REQUEST_1).length, Config.ServerAddress.getAddress(), Config.ServerAddress.getPort());
 
         byte[] responseBytes = new byte[4096];
         DatagramPacket responsePacket = new DatagramPacket(responseBytes, responseBytes.length);
@@ -108,7 +109,7 @@ class HandlerTest {
     void A2SPlayer() throws IOException {
         byte[] Response = joinArrays(ByteBufUtil.getBytes(Packets.A2S_PLAYER_REQUEST_HEADER), a2sChallenge);
         DatagramPacket queryPck = new DatagramPacket(Response, 0, Response.length,
-                Config.LocalServer.getAddress(), Config.LocalServer.getPort());
+                Config.ServerAddress.getAddress(), Config.ServerAddress.getPort());
 
         byte[] responseBytes = new byte[4096];
         DatagramPacket responsePacket = new DatagramPacket(responseBytes, responseBytes.length);
@@ -127,7 +128,7 @@ class HandlerTest {
     @Order(5)
     void A2SRulesChallenge() throws IOException {
         DatagramPacket queryPck = new DatagramPacket(ByteBufUtil.getBytes(Packets.A2S_RULES_CHALLENGE_REQUEST_1), 0,
-                ByteBufUtil.getBytes(Packets.A2S_RULES_CHALLENGE_REQUEST_1).length, Config.LocalServer.getAddress(), Config.LocalServer.getPort());
+                ByteBufUtil.getBytes(Packets.A2S_RULES_CHALLENGE_REQUEST_1).length, Config.ServerAddress.getAddress(), Config.ServerAddress.getPort());
 
         byte[] responseBytes = new byte[4096];
         DatagramPacket responsePacket = new DatagramPacket(responseBytes, responseBytes.length);
@@ -150,7 +151,7 @@ class HandlerTest {
     void A2SRules() throws IOException {
         byte[] Response = joinArrays(ByteBufUtil.getBytes(Packets.A2S_RULES_REQUEST_HEADER), a2sChallenge);
         DatagramPacket queryPck = new DatagramPacket(Response, 0, Response.length,
-                Config.LocalServer.getAddress(), Config.LocalServer.getPort());
+                Config.ServerAddress.getAddress(), Config.ServerAddress.getPort());
 
         byte[] responseBytes = new byte[4096];
         DatagramPacket responsePacket = new DatagramPacket(responseBytes, responseBytes.length);
